added IntervalDemo that shows how to use the scheduler, and how pause/resume works. Scheduler uses float instead of double (faster) Scheduler/CocosNode supports "[schedule .... ... interval:(float)]"
don't consume cpu while paused seems that it is not working... invistigate.


Former-commit-id: 04549bb4486a563b47cb60ce2ad9b8fdf256bdaf

+++ cocos2d/Particle.m
@@ -164,7 +164,7 @@
 	particle->size = size + sizeVar * RANDOM_FLOAT();	
 }
 
--(void) step: (double) dt
+-(void) step: (float) dt
 {
 	if( active ) {
 		float rate = 1.0 / emissionRate;
